# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Car Rent API
  version: 1.0.0
  description: |
    API для регистрации пользователей, бронирования автомобилей и работы со штрафами.
    Эта спецификация является первым артефактом проекта: реализация должна следовать ей.
servers:
  - url: http://127.0.0.1:8000
    description: Local development server
tags:
  - name: Users
    description: Регистрация пользователей
  - name: Bookings
    description: Бронирование автомобилей
  - name: Fines
    description: Штрафы пользователей
paths:
  /registration:
    post:
      tags:
        - Users
      operationId: registerUser
      summary: Зарегистрировать пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            example:
              license_number: 1234567890
              name: Анна
              surname: Иванова
              age: 21
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreated'
              example:
                user_id: 1
                message: Регистрация успешна
        '422':
          $ref: '#/components/responses/ValidationError'
  /bookings:
    post:
      tags:
        - Bookings
      operationId: createBooking
      summary: Забронировать автомобиль
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
            example:
              car_id: 10
              user_id: 1
              start_time: '2026-05-03T10:00:00Z'
              end_time: '2026-05-03T13:00:00Z'
      responses:
        '201':
          description: Бронирование создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingCreated'
              example:
                booking_id: 1
                car_id: 10
                user_id: 1
                total_price: 600
                status: confirmed
        '422':
          $ref: '#/components/responses/ValidationError'
  /bookings/{booking_id}:
    delete:
      tags:
        - Bookings
      operationId: cancelBooking
      summary: Отменить бронирование
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Бронирование отменено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingCancelled'
              example:
                booking_id: 1
                status: cancelled
                message: Бронирование успешно отменено
        '404':
          $ref: '#/components/responses/NotFound'
  /fines/{user_id}:
    get:
      tags:
        - Fines
      operationId: getUserFines
      summary: Посмотреть штрафы пользователя
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Информация о штрафах пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFines'
              example:
                user_id: 1
                has_fines: true
                fines_count: 1
                total_amount: 3000
                message: У пользователя есть штрафы
  /fines:
    post:
      tags:
        - Fines
      operationId: createFine
      summary: Начислить штраф пользователю
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FineCreate'
            example:
              user_id: 1
              amount: 3000
              reason: Превышение скорости
      responses:
        '201':
          description: Штраф начислен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FineCreated'
              example:
                fine_id: 1
                user_id: 1
                amount: 3000
                reason: Превышение скорости
                status: created
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  parameters:
    BookingId:
      name: booking_id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Идентификатор бронирования
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Идентификатор пользователя
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Бронь не найдена
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Некорректные данные запроса
  schemas:
    UserCreate:
      type: object
      required:
        - license_number
        - name
        - surname
        - age
      properties:
        license_number:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
        surname:
          type: string
          minLength: 1
        age:
          type: integer
          minimum: 18
        fine_count:
          type: integer
          minimum: 0
          default: 0
    UserCreated:
      type: object
      required:
        - user_id
        - message
      properties:
        user_id:
          type: integer
        message:
          type: string
    BookingCreate:
      type: object
      required:
        - car_id
        - user_id
        - start_time
        - end_time
      properties:
        car_id:
          type: integer
          minimum: 1
        user_id:
          type: integer
          minimum: 1
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
    BookingCreated:
      type: object
      required:
        - booking_id
        - car_id
        - user_id
        - total_price
        - status
      properties:
        booking_id:
          type: integer
        car_id:
          type: integer
        user_id:
          type: integer
        total_price:
          type: number
          format: float
        status:
          type: string
          enum:
            - confirmed
    BookingCancelled:
      type: object
      required:
        - booking_id
        - status
        - message
      properties:
        booking_id:
          type: integer
        status:
          type: string
          enum:
            - cancelled
        message:
          type: string
    FineCreate:
      type: object
      required:
        - user_id
        - amount
        - reason
      properties:
        user_id:
          type: integer
          minimum: 1
        amount:
          type: number
          format: float
          minimum: 1
        reason:
          type: string
          minLength: 1
    FineCreated:
      type: object
      required:
        - fine_id
        - user_id
        - amount
        - reason
        - status
      properties:
        fine_id:
          type: integer
        user_id:
          type: integer
        amount:
          type: number
          format: float
        reason:
          type: string
        status:
          type: string
          enum:
            - created
    UserFines:
      type: object
      required:
        - user_id
        - has_fines
        - fines_count
        - total_amount
        - message
      properties:
        user_id:
          type: integer
        has_fines:
          type: boolean
        fines_count:
          type: integer
        total_amount:
          type: number
          format: float
        message:
          type: string
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string